old-tests/submission/Infraware/Forms/contents/Forms/meter_max.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>meter_max</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form onsubmit="return false" name='input_form'>
		 <p>meter: <meter max=100 min=0 id='meter_id'></meter></p>
	</form>

	<script>
		
		var meter = document.getElementById("meter_id");
		meter.value = 50;
		var max = meter.max;
		meter.max = 200;

		if (typeof(meter.max) == "number") {
			if (max == 100 && meter.max == 200) {
				test(function() {
					assert_true(true, "max attribute is correct.");
				});
			} else {
				test(function() {
					assert_unreached("max attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("max attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
